+++ java/spark-1.6.0/sources.scala
@@ -860,6 +860,50 @@ object Sources {
     "    }\n" +
     "  }\n" +
     "}\n" +
+    "core_spark_apply_bundle_path <- function() {\n" +
+    "  file.path(tempdir(), \"packages.tar\")\n" +
+    "}\n" +
+    "\n" +
+    "#' Creates a bundle of dependencies required by \\code{spark_apply()}\n" +
+    "#'\n" +
+    "#' @keywords internal\n" +
+    "#' @export\n" +
+    "core_spark_apply_bundle <- function() {\n" +
+    "  packagesTar <- core_spark_apply_bundle_path()\n" +
+    "\n" +
+    "  args <- c(\"-cf\", packagesTar)\n" +
+    "  lapply(.libPaths(), function(e) {\n" +
+    "    args <<- c(args, \"-C\", e)\n" +
+    "    args <<- c(args, \".\")\n" +
+    "  })\n" +
+    "\n" +
+    "  if (!file.exists(packagesTar)) {\n" +
+    "    system2(\"tar\", args)\n" +
+    "  }\n" +
+    "\n" +
+    "  packagesTar\n" +
+    "}\n" +
+    "\n" +
+    "core_spark_apply_unbundle_path <- function() {\n" +
+    "  file.path(\"sparklyr-bundle\")\n" +
+    "}\n" +
+    "\n" +
+    "#' Extracts a bundle of dependencies required by \\code{spark_apply()}\n" +
+    "#'\n" +
+    "#' @param bundle_path Path to the bundle created using \\code{core_spark_apply_bundle()}\n" +
+    "#' @param base_path Base path to use while extracting bundles\n" +
+    "#'\n" +
+    "#' @keywords internal\n" +
+    "#' @export\n" +
+    "core_spark_apply_unbundle <- function(bundle_path, base_path) {\n" +
+    "  extractPath <- file.path(base_path, core_spark_apply_unbundle_path())\n" +
+    "\n" +
+    "  if (!dir.exists(extractPath)) dir.create(extractPath, recursive = TRUE)\n" +
+    "\n" +
+    "  system2(\"tar\", c(\"-xf\", bundle_path, \"-C\", extractPath))\n" +
+    "\n" +
+    "  extractPath\n" +
+    "}\n" +
     "core_get_package_function <- function(packageName, functionName) {\n" +
     "  if (packageName %in% rownames(installed.packages()) &&\n" +
     "      exists(functionName, envir = asNamespace(packageName)))\n" +
@@ -898,6 +942,31 @@ object Sources {
     "  )\n" +
     "\n" +
     "  worker_log(\"retrieved worker context\")\n" +
+    "\n" +
+    "  bundlePath <- worker_invoke(context, \"getBundlePath\")\n" +
+    "  if (nchar(bundlePath) > 0) {\n" +
+    "    worker_log(\"using bundle \", bundlePath)\n" +
+    "\n" +
+    "    if (file.exists(bundlePath)) {\n" +
+    "      worker_log(\"found local bundle exists, skipping extraction\")\n" +
+    "    }\n" +
+    "    else {\n" +
+    "      bundleName <- basename(bundlePath)\n" +
+    "\n" +
+    "      workerRootDir <- worker_invoke_static(sc, \"org.apache.spark.SparkFiles\", \"getRootDirectory\")\n" +
+    "      sparkBundlePath <- file.path(workerRootDir, bundleName)\n" +
+    "\n" +
+    "      if (!file.exists(sparkBundlePath)) {\n" +
+    "        stop(\"failed to find bundle under SparkFiles root directory\")\n" +
+    "      }\n" +
+    "\n" +
+    "      worker_log(\"updated .libPaths with bundle packages\")\n" +
+    "      unbundlePath <- core_spark_apply_unbundle(sparkBundlePath, workerRootDir)\n" +
+    "\n" +
+    "      .libPaths(unbundlePath)\n" +
+    "      worker_log(\"updated .libPaths with bundle packages\")\n" +
+    "    }\n" +
+    "  }\n" +
     "\n" +
     "  grouped_by <- worker_invoke(context, \"getGroupBy\")\n" +
     "  grouped <- !is.null(grouped_by)\n" +