+++ java/spark-1.6.0/worker.scala
@@ -14,7 +14,8 @@ class WorkerContext[T: ClassTag](
   closure: Array[Byte],
   columns: Array[String],
   groupBy: String,
-  closureRLang: Array[Byte]) {
+  closureRLang: Array[Byte],
+  bundlePath: String) {
 
   private var result: Array[T] = Array[T]()
 
@@ -67,6 +68,10 @@ class WorkerContext[T: ClassTag](
       lock.notify
     }
   }
+
+  def getBundlePath(): String = {
+    bundlePath
+  }
 }
 
 object WorkerHelper {
@@ -77,7 +82,8 @@ object WorkerHelper {
     port: Int,
     columns: Array[String],
     groupBy: String,
-    closureRLang: Array[Byte]): RDD[Row] = {
+    closureRLang: Array[Byte],
+    bundlePath: String): RDD[Row] = {
 
     val parent: RDD[Row] = rdd
     val computed: RDD[Row] = new WorkerRDD[Row](
@@ -87,7 +93,8 @@ object WorkerHelper {
       config,
       port,
       groupBy,
-      closureRLang)
+      closureRLang,
+      bundlePath)
 
     computed
   }