+++ NEWS.md
@@ -2,6 +2,9 @@
 
 ### Distributed R
 
+- Added `packages` parameter to `spark_apply()` to distribute packages
+  across worker nodes automatically.
+
 - Added `rlang` parameter to `spark_apply()`, to support by default
   generic closures provided by the `rlang` package.